refactor: enforce clip time fields and simplify annotation integer bounds
Agent-Logs-Url: https://github.com/lwy128384/jjj/sessions/22553720-6ddc-4de0-88d2-f7efa6931875

+++ train.py
@@ -163,8 +163,8 @@ def normalize_annotations(annotations):
             continue
         if e < s:
             s, e = e, s
-        s = float(math.floor(s))
-        e = float(math.ceil(e))
+        s = math.floor(s)
+        e = math.ceil(e)
         if e - s <= 1e-6:
             continue
         is_intf = bool(ann.get("is_interference", False))
